refactor: split main.ts into focused modules
Moves input reading into src/action-inputs.ts, the credential chain and
token exchange into src/auth.ts, the S3 client and its Yandex auth
middleware into src/s3-client.ts, file selection and upload into
src/upload.ts, and clearBucket into src/bucket.ts. main.ts keeps
orchestration only.

src/s3-client.ts is the one that mattered: the client construction, the two
removeByTag calls and the middleware insertion previously lived inline in
run() with no test of their own.

The 909-line main.test.ts splits to match, one file per module, with every
assertion moved unchanged.

Pure movement - the characterization snapshot is unchanged.
