+++ src/bucket.ts
@@ -1,0 +1,50 @@
+/**
+ * Bucket-wide operations.
+ *
+ * @module
+ */
+
+import { DeleteObjectsCommand, ListObjectsV2Command, S3Client } from '@aws-sdk/client-s3'
+import { info } from '@actions/core'
+
+/**
+ * Deletes every object in the bucket, one page of 1000 keys at a time.
+ *
+ * @param client - Configured S3 client
+ * @param bucket - Bucket name
+ */
+export async function clearBucket(client: S3Client, bucket: string): Promise<void> {
+    info('Clearing bucket')
+    const listCommand = new ListObjectsV2Command({
+        Bucket: bucket,
+        // The default and maximum number of keys returned is 1000.
+        MaxKeys: 1000
+    })
+
+    let isTruncated = true
+    let totalDeleted = 0
+
+    while (isTruncated) {
+        const { Contents, IsTruncated, NextContinuationToken } = await client.send(listCommand)
+
+        if (!Contents || Contents.length === 0) {
+            break
+        }
+
+        isTruncated = Boolean(IsTruncated)
+        listCommand.input.ContinuationToken = NextContinuationToken
+
+        const deleteCommand = new DeleteObjectsCommand({
+            Bucket: bucket,
+            Delete: {
+                Objects: Contents.map(c => ({ Key: c.Key }))
+            }
+        })
+
+        const { Deleted } = await client.send(deleteCommand)
+
+        totalDeleted += Deleted?.length ?? 0
+    }
+
+    info(`Deleted ${totalDeleted} objects from bucket ${bucket}`)
+}
